Improve test run (#748)
<!--
Thanks for taking precious time for making a PR.

Before creating a pull request, please make sure:
- Your PR solves one problem for which an issue exist and a solution has
been discussed
- You have read the guide for contributing
- See
https://github.com/beatlabs/patron/blob/master/README.md#how-to-contribute
- You signed all your commits (otherwise we won't be able to merge the
PR)
  - See https://github.com/beatlabs/patron/blob/master/SIGNYOURWORK.md
- You added unit tests for the new functionality
- You mention in the PR description which issue it is addressing, e.g.
"Resolves #123"
-->

## Which problem is this PR solving?

Resolves #747 
Resolves #746

+++ component/amqp/component.go
@@ -178,7 +178,7 @@ func (c *Component) processLoop(ctx context.Context, sub subscription) error {
 			if !ok {
 				return errors.New("subscription channel closed")
 			}
-			slog.Debug("processing message", slog.Int64("tag", int64(delivery.DeliveryTag)))
+			slog.Debug("processing message", slog.Uint64("tag", delivery.DeliveryTag))
 			observeReceivedMessageStats(ctx, c.queueCfg.queue, delivery.Timestamp)
 			c.processBatch(ctx, c.createMessage(ctx, delivery), btc)
 		case <-batchTimeout.C:
@@ -265,7 +265,7 @@ func (c *Component) createMessage(ctx context.Context, delivery amqp.Delivery) *
 func (c *Component) processBatch(ctx context.Context, msg *message, btc *batch) {
 	btc.append(msg)
 
-	if len(btc.messages) >= int(c.batchCfg.count) {
+	if uint(len(btc.messages)) >= c.batchCfg.count {
 		c.processAndResetBatch(ctx, btc)
 	}
 }
